[ADD] Refactored and normally realized session loading
