+++ app/data/module/main/OMS.class.ts
@@ -43,5 +43,6 @@ export default class OMS {
 
     OMSLanguage.checkLangsList();
     this.lang = new OMSLanguage(cfg.lang);
+    this.session = new OMSSession();
   }
 }
